// src/model/IProduct.ts
export type IProduct = {
  _id: string;
  images: any[];
  en: {
    title: string;
    midTitle: string;
    details: string;
    storage: string;
    direction: any[];
    types: any[];
  };
  ar: {
    title: string;
    midTitle: string;
    details: string;
    storage: string;
    direction: any[];
  };
};